style: treat query screen as a module
